Resolve review comments.
- Make clear_data_size an int.
- Use GL_RED_INTEGER when initializing variable.

+++ source/blender/gpu/GPU_storage_buffer.h
@@ -47,7 +47,7 @@ void GPU_storagebuf_clear_to_zero(GPUStorageBuf *ssbo);
  *
  * NOTE: clear_data_len must be in range 1 to 4 (inclusive).
  */
-void GPU_storagebuf_clear_uint(GPUStorageBuf *ssbo, uint32_t *clear_data, uint8_t clear_data_len);
+void GPU_storagebuf_clear_uint(GPUStorageBuf *ssbo, uint32_t *clear_data, int clear_data_len);
 
 /**
  * Clear the content of the buffer using the given #clear_data. #clear_data will be used as a
@@ -55,7 +55,7 @@ void GPU_storagebuf_clear_uint(GPUStorageBuf *ssbo, uint32_t *clear_data, uint8_
  *
  * NOTE: clear_data_len must be in range 1 to 4 (inclusive).
  */
-void GPU_storagebuf_clear_int(GPUStorageBuf *ssbo, int32_t *clear_data, uint8_t clear_data_len);
+void GPU_storagebuf_clear_int(GPUStorageBuf *ssbo, int32_t *clear_data, int clear_data_len);
 
 /**
  * Read back content of the buffer to CPU for inspection.
